Correctly handling VACTUAL 24bit signed integer
Negative readings of the signed 24 bit register VACTUAL resulted in wrong numbers because it didn't convert to 32 bit properly. The 24th sign bit needs to be "extended" to the unused 4th byte, which is why I'm ORing it with 0xFF000000 if the 24th bit is set (signifying a negative integer).
The same is most likely also true for PWM_SCALE_AUTO, a 9 bit signed int, but I haven't had the time to look into it.