- fixed SPI bus sharing between TFT, SD, Touch, etc in combination with TMC by properly initializing and then terminating the SPI usage
in the way that this library was previously written it was not easy to share the SPI bus between components because the SPI protocol was not properly terminated/cleaned-up across SPI sessions, causing conflicts for different SPI bus components. This library has been tested with Marlin 2.1.x bugfix on this repo: github.com/quiret/Marlin/

There is a Marlin pull request that would really appreciate this improvement: https://github.com/MarlinFirmware/Marlin/pull/24911

Please consider releasing another version of this library with the included fix so that the Marlin FW can be drastically improved! clobbering the global SPI object is not a good idea, especially on single-SPI boards...

Martin Turski, company owner of EirDev (turningtides@outlook.de)